Subject: Add Kotlin language supports
Explanatory Body:

This commit adds Kotlin language support to the project. It introduces
several new features and enhancements, including the ability to work
with strings and char sequences, extensions for JDK classes related to
files, IO, and threading, and various utility functions for working with
collections, math, and time.

Furthermore, this commit includes the addition of several libraries for
different purposes such as annotations, browsers, coroutines, DOM
manipulation, experimental features, native development, ranges,
reflection, streams, and more.

Please refer to the examples at
[here](https://play.kotlinlang.org/byExample/overview) for more details
on how to use Kotlin in this project.

Issue reference: N/A