feat: Add support for Kotlin language in the package
This change introduces support for the Kotlin programming language in
the package. Users can now import the 'kotlin.dart' file to utilize
Kotlin, along with other planned languages such as Java, C#, and Swift.
The package is designed to allow users to choose the language they want
to use in Dart. Additionally, longer examples have been added to the
`/example` folder. Contributing to the project is now possible by
opening an issue.

Resolves: #123